chore: move DebugScreen related stuff to the YAT.Debug namespace & remove IMonitorComponent interface
