+++ addons/yat/src/Debug/looking_at_info/LookingAtInfo.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://hvc8a2qwximn"]
 
-[ext_resource type="Script" uid="uid://bgpyf6saegeop" path="res://addons/yat/src/scenes/debug_screen/items/looking_at_info/LookingAtInfo.cs" id="1_lxy07"]
+[ext_resource type="Script" uid="uid://bgpyf6saegeop" path="res://addons/yat/src/Debug/items/looking_at_info/LookingAtInfo.cs" id="1_lxy07"]
 
 [node name="LookingAtInfo" type="PanelContainer"]
 script = ExtResource("1_lxy07")
