+++ addons/yat/src/Debug/scene_objects_info/SceneObjectsInfo.cs
@@ -1,8 +1,7 @@
 using Godot;
 using YAT.Attributes;
-using YAT.Interfaces;
 
-namespace YAT.Scenes;
+namespace YAT.Debug;
 
 [Title("SceneObjects")]
 public partial class SceneObjectsInfo : PanelContainer, IDebugScreenItem
