+++ addons/yat/src/Debug/scene_objects_info/SceneObjectsInfo.cs.uid
