+++ addons/yat/src/Debug/scene_objects_info/SceneObjectsInfo.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://lscv6c8lgnyh"]
 
-[ext_resource type="Script" uid="uid://dwpa5b2p34jvi" path="res://addons/yat/src/scenes/debug_screen/items/scene_objects_info/SceneObjectsInfo.cs" id="1_wkyk1"]
+[ext_resource type="Script" uid="uid://dwpa5b2p34jvi" path="res://addons/yat/src/Debug/items/scene_objects_info/SceneObjectsInfo.cs" id="1_wkyk1"]
 
 [node name="SceneObjectsInfo" type="PanelContainer"]
 script = ExtResource("1_wkyk1")
