--- addons/yat/src/Net/EConnectionStatus.uid
@@ -1,1 +1,0 @@
-uid://bmnc31c3fxby4
