--- addons/yat/src/interfaces/IMonitorComponent.cs
@@ -1,8 +1,0 @@
-namespace YAT.Interfaces;
-
-public interface IMonitorComponent
-{
-    public bool UseColors { get; set; }
-
-    public void Update();
-}
