--- addons/yat/src/interfaces/IMonitorComponent.cs.uid
@@ -1,1 +1,0 @@
-uid://chstegxnmha7k
